requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
gitpython>=3.1.0
